backend/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-multipart>=0.0.6
pandas>=2.2.0
numpy>=1.26.0
scikit-learn>=1.3.2
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-dotenv>=1.0.0
supabase>=2.0.3
openpyxl>=3.1.2
reportlab>=4.0.7
scipy>=1.11.0
xgboost>=2.0.0
joblib>=1.3.0
sqlalchemy>=2.0.23
psycopg2-binary>=2.9.9
alembic>=1.12.1
matplotlib>=3.8.0
seaborn>=0.13.0